Simplify Power integral narrowing conversion evaluators
The Power codegen's evaluators for performing integral narrowing
conversions (e.g. i2s) have been significantly simplified to remove
unneeded handling. Previously, some of these evaluators would perform
sign- and zero- extension, which seems to be a remnant of a time when
the upper bits of sub-integer values in registers were not considered to
be undefined.

Signed-off-by: Ben Thomas <ben@benthomas.ca>