Fix s2l not sign-extending correctly
It turns out that the evaluator for s2l was incorrectly sign-extending
its source register as an integer rather than as a short. This was
previously masked by the fact that many evaluators would unnecessarily
sign-extend their results. In general, the upper bits of a register
containing a short are undefined, so it is incorrect to assume it has
been sign-extended.

Signed-off-by: Ben Thomas <ben@benthomas.ca>