Modifications for Off-Heap Technology for Large Arrays
Modifications required to complete the introduction of an optimized memory
management scheme for large arrays in a region-based garbage collector.
More specifically, this is a new heap scheme and large
object organization (in-heap and off-heap), an
enhanced method for allocating large objects (proxy
objects), and an optimized technique of accessing
large arrays during runtime (JIT optimizations).

Related to: #14005

Signed-off-by: Jonathan Oommen jon.oommen@gmail.com