Add declared class to Symbol, like a signature but as a class pointer
...and consult it when constraining aloadi in VP.

With this declared class (once it is populated appropriately) VP can
determine the type of the result of aloadi of fabricated field shadows.

We don't store signatures, but rather get them based on the owning
method and the constant pool index, so in particular after fabricating a
field shadow we lose track of the signature. Furthermore, even if we had
a signature in that case, we wouldn't necessarily have any owning method
that would be appropriate to use for looking up the class by name.

While creating a fabricated field shadow, however, the compiler will
have access to the definition of the containing type, so it will be able
to look up the correct class and remember it.

The aloadi VP handler now ignores the signature in favour of the
declared class whenever the latter is present. It's fine to set the
declared class for regular (i.e. non-fabricated) field shadows as well,
in which case we can simply avoid repeated lookups. Some logic using the
signature in this VP handler has been changed to use the class pointer
classBlock instead, since it only ran when classBlock was available
anyway. The other use of the signature is deleted because it consulted
field info from class lookahead, which is broken and not enabled.

The declared class is printed along with the DataType in the verbose
incremental symref table output. Additionally, named shadows now print
"Named Shadow" instead of repeating the name a second time.