Expand bytecode offset variables to 32bit
A user experienced a compile time crash in setHandlerInfoWithOutBCInfo()
because a bytecode index to a catch handler is above the 2^16th limit
that normally applies for Java methods. This limit can be exceeded in
rare cases after the JVM expands the method size on load. A good example
of this would be when the VM inlines a "jsr" bytecode. Because the
handler index is larger then 2^16th the offset is truncated when it is
placed into the 16bit integer that we were using to record the offset.
This fix will expand the size of variables used to hold bytecode offsets
to be 32bits.

Signed-off-by: Kevin Langman <langman@ca.ibm.com>