Ignore errors, delay and cache the result of isRunningInContainer
In #6432, OMR port library started throwing an error if
isRunningInContainer failed. isRunningInContainer can fail if /proc
is mounted with the hidepid=2 setting on Linux (#7021). This prevents
a JVM user to start. Before #6432, no error was returned if
isRunningInContainer failed; a user was completely unaware of this
failure; this behaviour can lead to performance issues if the process
is running in a container; but no functional issues will be seen. The
new behaviour will not throw an error if isRunningInContainer fails,
but will issue a warning message to highlight the potential performance
impact.

Currently, isRunningInContainer is run from omrsysinfo_startup.
Neither the trace engine nor NLS messages are enabled at this
point. If there is an error in isRunningInContainer, no tracepoint
or NLS message will work inside isRunningInContainer.

Invocation of isRunningInContainer is delayed to first-use. In OpenJ9,
the first-use still happens before the trace engine is initialized,
but it happens after the NLS messages are enabled. A new NLS message
has been added in eclipse-openj9/openj9#17560, which will show up as
a warning when isRunningInContainer fails and highlight the potential
performance impact.

The result of isRunningInContainer is cached and updated via an atomic
operation to enforce data consistency. The caching helps to improve
performance when isRunningInContainer is repeatedly invoked.

Four new states are introduced for PPG_isRunningInContainer to support
the new changes:
- OMRPORT_RUNNING_IN_CONTAINER_UNINITIALIZED: evaluate the result of
  isRunningInContainer.
- OMRPORT_RUNNING_IN_CONTAINER_TRUE: inside a container.
- OMRPORT_RUNNING_IN_CONTAINER_FALSE: not in a container.
- OMRPORT_RUNNING_IN_CONTAINER_ERROR: an error is encountered while
  evaluating isRunningInContainer.

Related: #7021

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>