Modifications for Off-Heap Technology for Large Arrays
Modifications required to complete the introduction of an optimized
memory management scheme for large arrays in a region-based garbage
collector.
More specifically, this is a new heap scheme and large
object organization (in-heap and off-heap), an
enhanced method for allocating large objects (proxy
objects), and an optimized technique of accessing
large arrays during runtime (JIT optimizations).

 - new define OMR_GC_SPARSE_HEAP_ALLOCATION
 - set OMR_GC_SPARSE_HEAP_ALLOCATION ON
 - obsolete Trc_MM_double_map_Entry, once all usages are updated.
 - replace J9VM_GC_ENABLE_SPARSE_HEAP_ALLOCATION with
   OMR_GC_SPARSE_HEAP_ALLOCATION.
 - empty method offsetOfContiguousDataAddrField() in ObjectModel
 - empty method generateArrayElementShiftAmountTrees() in TransformUtil
 - empty method isOffHeapAllocationEnabled() in ObjectModel

Signed-off-by: hulin <linhu@ca.ibm.com>