Do not fail retrieveLinuxCgroupMemoryStats if swappiness is unavailable
`retrieveLinuxCgroupMemoryStats()` used to return an error failure
if `swappiness` cannot be determined. However, `swappiness` is not
available for Cgroup v2. Moreover, `swappiness` is not a critical
factor, while knowing the memory limits for your container typically
is.
This commit changes `retrieveLinuxCgroupMemoryStats()` to return
a success code if the only information that could not be retrieved
is the `swappiness`. The value of the returned `swappiness` is set
to `OMRPORT_MEMINFO_NOT_AVAILABLE` in this case.

Fixes: #7453

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>