Allow pre-eval escaping conversion nodes of evaluated array-access
preEvaluateEscapingNodesForSpineCheck gets called to pre-evaluate a
future used commoned node before branching out of mainline code.
It restricts pre-evaluating conversions with array-access children for
correctness of access. But if array-access is already evaluated this
restriction is non-beneficial and might prevent a needed pre-evaluation
into the mainline code for a future used commoned node.